Create README.md
This is an AI-powered bot designed to provide emotional support and assistance to individuals struggling with mental health issues. It can help individuals access mental health resources offer guidance and support. With the integration of Language translation this chatbot will be very efficient as it will be able to break the language barriers. The creation of a chatbot capable of language translation holds transformative potential acting as a catalyst in overcoming language barriers for effective communication and information exchange. Its impact spans diverse sectors including: healthcare commerce and governance etc. offering a versatile solution to bridge linguistic gaps.